implemented new pattern for enums
