+++ tests/CMakeLists.txt
@@ -16,10 +16,10 @@ set_target_properties(enki_tests PROPERTIES
 
 target_sources(enki_tests PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/regression_tests/arithmetic_serdes.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/regression_tests/enum_serdes.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/regression_tests/array_serdes.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/regression_tests/base_engine_composite_tests.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/regression_tests/base_engine_custom_type_serdes.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/regression_tests/base_engine_enum_serdes.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/regression_tests/base_engine_range_constructible_serdes.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/regression_tests/base_engine_requirements.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/regression_tests/base_engine_tuple_like_serdes.cpp
